.cdk/iac/requirements.txt
aws-cdk.core
aws-cdk.aws-autoscaling
aws-cdk.aws-ec2
aws-cdk.aws-ssm
aws-cdk.aws_ssm
aws-cdk.aws_iam
